Move bib fetch functions to server dir
