+++ src/server/api/bib.ts
@@ -1,16 +1,16 @@
-import type { BibParams, BibResponse } from "../../../src/types/bibTypes"
+import type { BibParams, BibResponse } from "../../types/bibTypes"
 import {
   getBibQuery,
   isNyplBibID,
   standardizeBibId,
-} from "../../../src/utils/bibUtils"
-import nyplApiClient from "../../../src/server/nyplApiClient"
+} from "../../utils/bibUtils"
+import nyplApiClient from "../nyplApiClient"
 import {
   DISCOVERY_API_NAME,
   DISCOVERY_API_SEARCH_ROUTE,
   SHEP_HTTP_TIMEOUT,
-} from "../../../src/config/constants"
-import { appConfig } from "../../../src/config/config"
+} from "../../config/constants"
+import { appConfig } from "../../config/config"
 
 export async function fetchBib(
   id: string,
